+++ app/src/androidTest/java/com/github/se/gomeet/screens/TrendsScreen.kt
@@ -7,8 +7,7 @@ import io.github.kakaocup.compose.node.element.KNode
 
 class TrendsScreen(semanticsProvider: SemanticsNodeInteractionsProvider) :
     ComposeScreen<TrendsScreen>(
-        semanticsProvider = semanticsProvider,
-        viewBuilderAction = { hasTestTag("TrendsScreen") }) {
+        semanticsProvider = semanticsProvider, viewBuilderAction = { hasTestTag("TrendsScreen") }) {
 
-    val createButton: KNode = child { hasTestTag(Route.CREATE) }
+  val createButton: KNode = child { hasTestTag(Route.CREATE) }
 }
