+++ app/src/main/java/com/github/se/gomeet/MainActivity.kt
@@ -49,17 +49,21 @@ class MainActivity : ComponentActivity() {
                   onSignInSuccess = { userId ->
                     userIdState.value = userId
                     NavigationActions(nav)
-                        .navigateTo(TOP_LEVEL_DESTINATIONS.first{it.route == Route.CREATE}, clearBackStack = true)
+                        .navigateTo(
+                            TOP_LEVEL_DESTINATIONS.first { it.route == Route.CREATE },
+                            clearBackStack = true)
                   })
             }
             composable(Route.LOGIN) {
               LoginScreen(viewModel) {
-                NavigationActions(nav).navigateTo(TOP_LEVEL_DESTINATIONS.first{it.route == Route.CREATE})
+                NavigationActions(nav)
+                    .navigateTo(TOP_LEVEL_DESTINATIONS.first { it.route == Route.CREATE })
               }
             }
             composable(Route.REGISTER) {
               RegisterScreen(viewModel) {
-                NavigationActions(nav).navigateTo(TOP_LEVEL_DESTINATIONS.first{it.route == Route.CREATE})
+                NavigationActions(nav)
+                    .navigateTo(TOP_LEVEL_DESTINATIONS.first { it.route == Route.CREATE })
               }
             }
             composable(Route.EXPLORE) { Explore(navAction, EventViewModel()) }
